testing/web-platform/tests/css/css-typed-om/stylevalue-subclasses/numeric-objects/numeric-typing.tentative.html

<!doctype html>
<meta charset="utf-8">
<title>Numeric value typing tests</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#numeric-typing">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

const gBaseTypes = [
  {
    baseType: 'number',
    units: ['number']
  },
  {
    baseType: 'percent',
    units: ['percent']
  },
  {
    baseType: 'length',
    units: ['em', 'ex', 'ch', 'ic', 'rem', 'lh', 'rlh', 'vw', 'vh', 'vi', 'vb', 'vmin', 'vmax', 'cm', 'mm', 'q', 'in', 'pt', 'pc', 'px']
  },
  {
    baseType: 'time',
    units: ['s', 'ms']
  },
  {
    baseType: 'frequency',
    units: ['Hz', 'kHz']
  },
  {
    baseType: 'resolution',
    units: ['dpi', 'dpcm', 'dppx']
  },
  {
    baseType: 'flex',
    units: ['fr']
  },
];

for (const {baseType: baseType1, units: units1} of gBaseTypes) {
  for (const {baseType: baseType2, units: units2} of gBaseTypes) {
    if (baseType1 == baseType2) {
      test(() => {
        const value1 = new CSSUnitValue(0, units1[0]);
        const value2 = new CSSUnitValue(0, units2[0]);
        new CSSMathSum(value1, value2);
      }, 'Adding two ' + baseType1 + ' values does not throw error');
    } else {
      if ((baseType1 == 'percent' || baseType2 == 'percent') &&
          (baseType1 != 'number' && baseType2 != 'number')) {
        test(() => {
          const value1 = new CSSUnitValue(0, units1[0]);
          const value2 = new CSSUnitValue(0, units2[0]);
          new CSSMathSum(value1, value2);
        }, 'Adding a ' + baseType1 + ' with a ' + baseType2 + ' does not throw error');
      } else {
        test(() => {
          const value1 = new CSSUnitValue(0, units1[0]);
          const value2 = new CSSUnitValue(0, units2[0]);
          assert_throws(new TypeError(), () => new CSSMathSum(value1, value2));
        }, 'Adding a ' + baseType1 + ' with a ' + baseType2 + ' throws TypeError');
      }
    }
  }
}

</script>
